@@ -5,14 +5,10 @@
 // Copyright (C) Jeff Hansen 2015. All rights reserved.
 
 using System.Collections.Generic;
-using System.Text;
-
 using Jeffijoe.MessageFormat.Formatting;
 using Jeffijoe.MessageFormat.Formatting.Formatters;
 using Jeffijoe.MessageFormat.Parsing;
-
-using Moq;
-
+using Jeffijoe.MessageFormat.Tests.TestHelpers;
 using Xunit;
 
 namespace Jeffijoe.MessageFormat.Tests.Formatting.Formatters
@@ -27,12 +23,13 @@ namespace Jeffijoe.MessageFormat.Tests.Formatting.Formatters
         /// <summary>
         /// Gets the format_tests.
         /// </summary>
-        public static IEnumerable<object[]> Format_tests
+        public static IEnumerable<object[]> FormatTests
         {
             get
             {
                 yield return new object[] { "male {he said} female {she said} other {they said}", "male", "he said" };
-                yield return new object[] { "male {he said} female {she said} other {they said}", "female", "she said" };
+                yield return new object[]
+                    { "male {he said} female {she said} other {they said}", "female", "she said" };
                 yield return new object[] { "male {he said} female {she said} other {they said}", "dawg", "they said" };
             }
         }
@@ -54,21 +51,40 @@ namespace Jeffijoe.MessageFormat.Tests.Formatting.Formatters
         /// The expected block.
         /// </param>
         [Theory]
-        [MemberData(nameof(Format_tests))]
+        [MemberData(nameof(FormatTests))]
         public void Format(string formatterArgs, string gender, string expectedBlock)
         {
             var subject = new SelectFormatter();
-            var messageFormatterMock = new Mock<IMessageFormatter>();
-            messageFormatterMock.Setup(x => x.FormatMessage(It.IsAny<string>(), It.IsAny<Dictionary<string, object?>>()))
-                                .Returns((string input, Dictionary<string, object> a) => input);
+            var messageFormatter = new FakeMessageFormatter();
             var req = new FormatterRequest(
-                new Literal(1, 1, 1, 1, ""), 
-                "gender", 
-                "select", 
+                new Literal(1, 1, 1, 1, ""),
+                "gender",
+                "select",
                 formatterArgs);
             var args = new Dictionary<string, object?> { { "gender", gender } };
-            var result = subject.Format("en", req, args, gender, messageFormatterMock.Object);
+            var result = subject.Format("en", req, args, gender, messageFormatter);
             Assert.Equal(expectedBlock, result);
+        }
+
+        /// <summary>
+        /// Verifies that format throws when no other option is given.
+        /// </summary>
+        [Fact]
+        public void VerifyFormatThrowsWhenNoOtherOptionIsGiven()
+        {
+            var subject = new SelectFormatter();
+            var messageFormatter = new FakeMessageFormatter();
+            var req = new FormatterRequest(
+                new Literal(1, 1, 1, 1, ""),
+                "gender",
+                "select",
+                "male {he} female{she}");
+            var args = new Dictionary<string, object?> { { "gender", "non-binary" } };
+
+            Assert.Throws<MessageFormatterException>(() =>
+            {
+                subject.Format("en", req, args, "non-binary", messageFormatter);
+            });
         }
 
         #endregion
