@@ -5,10 +5,8 @@
 // Copyright (C) Jeff Hansen 2015. All rights reserved.
 
 using System.Collections.Generic;
-
 using Jeffijoe.MessageFormat.Formatting;
 using Jeffijoe.MessageFormat.Tests.TestHelpers;
-
 using Xunit;
 using Xunit.Abstractions;
 
@@ -119,6 +117,20 @@ namespace Jeffijoe.MessageFormat.Tests
                         new Dictionary<string, object?> { { "num", 2 } },
                         "#{2#2}"
                     };
+                yield return
+                    new object[]
+                    {
+                        "'''{'''",
+                        new Dictionary<string, object?>(),
+                        "'{'"
+                    };
+                // yield return
+                //     new object[]
+                //     {
+                //         "{num, plural, =1 {1} other {'''{'''#'''}'''}}",
+                //         new Dictionary<string, object?> { { "num", 2 } },
+                //         "'{'2'}'"
+                //     };
             }
         }
 
@@ -208,7 +220,8 @@ namespace Jeffijoe.MessageFormat.Tests
                     new object[]
                     {
                         Case2,
-                        new Dictionary<string, object?> { { "gender", "female" }, { "name", "Amanda" }, { "count", 1 } },
+                        new Dictionary<string, object?>
+                            { { "gender", "female" }, { "name", "Amanda" }, { "count", 1 } },
                         "She - {Amanda} - said: You have just one notification. Have a nice day!"
                     };
                 yield return
@@ -271,7 +284,8 @@ namespace Jeffijoe.MessageFormat.Tests
                     new object[]
                     {
                         Case5,
-                        new Dictionary<string, object?> { { "count", 42 }, { "gender", "female" }, { "genitals", 102 } },
+                        new Dictionary<string, object?>
+                            { { "count", 42 }, { "gender", "female" }, { "genitals", 102 } },
                         "She (who has the freakish amount of 102 boobies) said: You have a universal amount of notifications. Have a nice day!"
                     };
                 yield return
@@ -333,6 +347,13 @@ namespace Jeffijoe.MessageFormat.Tests
                         Case6,
                         new Dictionary<string, object?> { { "count", 3 } },
                         "You and 2 others added this to their profiles."
+                    };
+                yield return
+                    new object[]
+                    {
+                        "{ count, plural, one {1 thing} other {# things} }",
+                        new Dictionary<string, object?> { { "count", 2 } },
+                        "2 things"
                     };
             }
         }
@@ -581,8 +602,9 @@ namespace Jeffijoe.MessageFormat.Tests
             }
 
             {
-                var mf = new MessageFormatter(true, "en");
-                mf.Pluralizers!["en"] = n => {
+                var mf = new MessageFormatter(useCache: true, locale: "en");
+                mf.Pluralizers!["en"] = n =>
+                {
                     // ´n´ is the number being pluralized.
                     // ReSharper disable once CompareOfFloatsByEqualityOperator
                     if (n == 0)
@@ -590,6 +612,7 @@ namespace Jeffijoe.MessageFormat.Tests
                         return "zero";
                     }
 
+                    // ReSharper disable once CompareOfFloatsByEqualityOperator
                     if (n == 1)
                     {
                         return "one";
