@@ -4,13 +4,7 @@
 // Author: Jeff Hansen <jeff@jeffijoe.com>
 // Copyright (C) Jeff Hansen 2015. All rights reserved.
 
-using System.Collections.Generic;
-
-using Jeffijoe.MessageFormat.Formatting;
-using Jeffijoe.MessageFormat.Tests.TestHelpers;
-
 using Xunit;
-using Xunit.Abstractions;
 
 namespace Jeffijoe.MessageFormat.Tests
 {
@@ -19,30 +13,8 @@ namespace Jeffijoe.MessageFormat.Tests
     /// </summary>
     public class MessageFormatterIssues
     {
-        #region Fields
-
-        /// <summary>
-        /// The output helper.
-        /// </summary>
-        private readonly ITestOutputHelper outputHelper;
-
-        #endregion
-
-        #region Constructors and Destructors
-
-        /// <summary>
-        /// Ctor.
-        /// </summary>
-        /// <param name="outputHelper"></param>
-        public MessageFormatterIssues(ITestOutputHelper outputHelper)
-        {
-            this.outputHelper = outputHelper;
-        }
-
-        #endregion
-
         [Fact]
-        public void Issue13()
+        public void Issue13_Bad_escaping_on_pound_symbol()
         {
             string plural = @"{num_guests, plural, offset:1, other {# {host} invites # people to their party.}}";
             string broken = @"{num_guests, plural, offset:1, other {{host} invites # people to their party.}}";
@@ -51,6 +23,18 @@ namespace Jeffijoe.MessageFormat.Tests
             var vars = new { num_guests = "5", host = "Mary" };
             Assert.Equal("Mary invites 4 people to their party.", mf.FormatMessage(broken, vars));
             Assert.Equal("4 Mary invites 4 people to their party.", mf.FormatMessage(plural, vars));
+        }
+        
+        [Fact]
+        public void Issue27_WhiteSpace_in_identifiers_is_ignored()
+        {
+            var subject = new MessageFormatter(false);
+            var result = subject.FormatMessage("{ count, plural , one {1 thing} other {# things} }", new
+            {
+                count = 2
+            });
+
+            Assert.Equal("2 things", result);
         }
     }
 }
