@@ -1,7 +1,8 @@
 ﻿namespace Jeffijoe.MessageFormat.Formatting.Formatters
 {
+    [System.Diagnostics.CodeAnalysis.ExcludeFromCodeCoverage]
     internal static partial class PluralRulesMetadata
     {
-        public static partial bool TryGetRuleByLocale(string locale, out ContextPluralizer pluralizer);
+        public static partial bool TryGetRuleByLocale(string locale, out ContextPluralizer contextPluralizer);
     }
 }
